backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
numpy==1.23.5
pandas==1.5.3
opencv-python==4.7.0.72
dlib==19.24.0
matplotlib==3.7.1
pillow==9.4.0
gunicorn==20.1.0
pydicom==2.3.1